Rename `workouts-table-types.ts` file to `types.ts` and update imports
