+++ src/components/workouts-table/workouts-table-toolbar/WorkoutsTableToolbar.tsx
@@ -5,7 +5,7 @@ import { SetStoreFunction } from 'solid-js/store'
 
 import { WorkoutsTableDialog } from '~/components/workouts-table'
 
-import type { Workout } from '~/components/workouts-table/workouts-table-types'
+import type { Workout } from '~/components/workouts-table/types'
 
 export default function WorkoutsTableToolbar(props: {
   setWorkouts: SetStoreFunction<Workout[]>
